Move decision-related utils to their own util
This prevents some issues with dependency cycles, and keeps things a tad
more tidy
