+++ server/utils/assessments/utils.test.ts
@@ -4,7 +4,6 @@ import {
   allocatedTableRows,
   allocationLink,
   allocationSummary,
-  applicationAccepted,
   arriveDateAsTimestamp,
   assessmentLink,
   assessmentSections,
@@ -18,7 +17,6 @@ import {
   daysSinceInfoRequest,
   daysSinceReceived,
   daysUntilDue,
-  decisionFromAssessment,
   formatDays,
   formatDaysUntilDueWithWarning,
   formattedArrivalDate,
@@ -581,48 +579,6 @@ describe('utils', () => {
           category: 'prison-information',
         })}">View additional prison information</a></p>`,
       )
-    })
-  })
-
-  describe('decisionFromAssessment', () => {
-    it('returns the decision from the assessment if it exists', () => {
-      const assessment = assessmentFactory.build({
-        data: { 'make-a-decision': { 'make-a-decision': { decision: 'the decision' } } },
-      })
-      expect(decisionFromAssessment(assessment)).toEqual('the decision')
-    })
-
-    it('returns an empty string of the decision doesnt exist', () => {
-      const assessment = assessmentFactory.build({ data: {} })
-      expect(decisionFromAssessment(assessment)).toEqual('')
-    })
-  })
-
-  describe('applicationAccepted', () => {
-    it('returns true if the assessment has either of the two decisions which accept an applicaiton', () => {
-      const acceptedAssessment1 = assessmentFactory.build({
-        data: { 'make-a-decision': { 'make-a-decision': { decision: 'releaseDate' } } },
-      })
-      const acceptedAssessment2 = assessmentFactory.build({
-        data: { 'make-a-decision': { 'make-a-decision': { decision: 'releaseDate' } } },
-      })
-
-      expect(applicationAccepted(acceptedAssessment1)).toBe(true)
-      expect(applicationAccepted(acceptedAssessment2)).toBe(true)
-    })
-
-    it('returns false if the assessment has any other decision', () => {
-      const rejectedAssessment = assessmentFactory.build({
-        data: { 'make-a-decision': { 'make-a-decision': { decision: 'reject' } } },
-      })
-
-      expect(applicationAccepted(rejectedAssessment)).toBe(false)
-    })
-
-    it('returns false if the assessment has no decision', () => {
-      const rejectedAssessment = assessmentFactory.build()
-
-      expect(applicationAccepted(rejectedAssessment)).toBe(false)
     })
   })
